Improve nonce security
Add nonce endpoint which generates an unique nonce when read.
This nonce is valid for 30 seconds.

The user must give this nonce and the signature of it when logging in.
This fixes replay attacks and generating signatures in the future.
See #7 for more info.

This adds an extra `secure_nonce` option to the configuration, which by
default is true. If set to false you can still use the timebased nonces.

We also add a periodicFunc function which remove all old unused nonces
every minute.

Dependencies are also updated to the latest releases.

Closes #7